isr-system/requirements.txt
# ==================================================
# requirements.txt — Pin every dep for reproducibility
# Install: pip install -r requirements.txt
# ==================================================

# --- Core ---
torch==2.1.0
torchvision==0.16.0

# --- Data & Image Processing ---
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# --- Quality Metrics ---
piq==0.8.0

# --- Optimisation ---
onnx==1.15.0
onnxruntime-gpu==1.16.0
# TensorRT: install separately via NVIDIA wheel
# pip install tensorrt==8.6.1 --extra-index-url https://pypi.ngc.nvidia.com

# --- Experiment Tracking ---
mlflow==2.8.1

# --- Data Versioning ---
dvc==3.30.1

# --- Config & Validation ---
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dotenv==1.0.0

# --- API & Deployment ---
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# --- Monitoring ---
prometheus-client==0.18.0

# --- Testing ---
pytest==7.4.3
pytest-benchmark==4.0.0
pytest-cov==4.1.0

# --- Code Quality ---
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
